Refactoring code layout
