+++ communication/comm.h
