+++ general/dim.h
