+++ samplers/algorithm/algo_aitken.h
