+++ samplers/spatial/sampler_gauss.h
