+++ samplers/spatial/sampler_mov_avg.h
