+++ samplers/spatial/sampler_nn.h
