+++ samplers/spatial/sampler_pseudo_n2_linear.h
