+++ samplers/spatial/sampler_shepard_quintic.h
