+++ samplers/spatial/sampler_sph_quintic.h
