+++ samplers/spatial/sampler_sum_quintic.h
