+++ samplers/temporal/temporal_sampler_gauss.h
