+++ samplers/temporal/temporal_sampler_null.h
